TsumoControllerをpublicなクラスに変更
